MCP_tools/requirements.txt
crewai
fastapi
uvicorn
pydantic
python-dotenv
fastmcp
google-auth
google-auth-oauthlib
google-api-python-client
